content/html/content/test/test_bug549475.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=549475
-->
<head>
  <title>Test for Bug 549475</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=549475">Mozilla Bug 549475</a>
<p id="display"></p>
<pre id="test">
<div id='content'>
  <form>
  </form>
</div>
<script type="application/javascript">

/** Test for Bug 549475 **/

// We are excluding "file" because it's too different from the other types.
// And it has no sanitizing algorithm.
var inputTypes =
[
  "text", "password", "search", "tel", "hidden", "checkbox", "radio",
  "submit", "image", "reset", "button", "email", "url", "number",
];

var todoTypes =
[
  "range", "color",
  "date", "month", "week", "time", "datetime", "datetime-local",
];

var valueModeValue =
[
  "text", "search", "url", "tel", "email", "password", "date", "datetime",
  "month", "week", "time", "datetime-local", "number", "range", "color",
];

function sanitizeValue(aType, aValue)
{
  switch (aType) {
    case "text":
    case "password":
    case "search":
    case "tel":
      return aValue.replace(/[\n\r]/g, "");
    case "url":
    case "email":
      return aValue.replace(/[\n\r]/g, "").replace(/^\s+|\s+$/g, "");
    case "date":
    case "month":
    case "week":
    case "time":
    case "datetime":
    case "datetime-local":
      // TODO: write the sanitize algorithm.
      return "";
    case "number":
      return (parseFloat(aValue) + "" === aValue) ? aValue : "";
    case "range":
      // TODO: write the sanitize algorithm.
      return "";
    case "color":
      // TODO: write the sanitize algorithm.
      return "";
    default:
      return aValue;
  }
}

function checkSanitizing(element)
{
  var testData =
  [
    // For text, password, search, tel, email:
    "\n\rfoo\n\r",
    "foo\n\rbar",
    "  foo  ",
    "  foo\n\r  bar  ",
    // For url:
    "\r\n foobar    \n\r",
    // For number:
    "42",
    "13.37",
    "1.234567898765432",
  ];

  for (value of testData) {
    element.setAttribute('value', value);
    is(element.value, sanitizeValue(type, value),
       "The value has not been correctly sanitized for type=" + type);
    is(element.getAttribute('value'), value,
       "The content value should not have been sanitized");

    if (type in valueModeValue) {
      element.setAttribute('value', 'tulip');
      element.value = value;
      is(element.value, sanitizeValue(type, value),
         "The value has not been correctly sanitized for type=" + type);
      is(element.getAttribute('value'), 'tulip',
         "The content value should not have been sanitized");
    }

    element.setAttribute('value', '');
    form.reset();
    element.type = 'checkbox'; // We know this type has no sanitizing algorithm.
    element.setAttribute('value', value);
    is(element.value, value, "The value should not have been sanitized");
    element.type = type;
    is(element.value, sanitizeValue(type, value),
       "The value has not been correctly sanitized for type=" + type);
    is(element.getAttribute('value'), value,
       "The content value should not have been sanitized");

    element.setAttribute('value', '');
    form.reset();
    element.setAttribute('value', value);
    form.reset();
    is(element.value, sanitizeValue(type, value),
       "The value has not been correctly sanitized for type=" + type);
    is(element.getAttribute('value'), value,
       "The content value should not have been sanitized");

    // Cleaning-up.
    element.setAttribute('value', '');
    form.reset();
  }
}

var pref = SpecialPowers.getBoolPref("dom.experimental_forms");
SpecialPowers.setBoolPref("dom.experimental_forms", true);
for (type of inputTypes) {
  var form = document.forms[0];
  var element = document.createElement("input");
  element.style.display = "none";
  element.type = type;
  form.appendChild(element);

  checkSanitizing(element); // no frame, no editor

  element.style.display = "";
  checkSanitizing(element); // frame, no editor

  element.focus();
  element.blur();
  checkSanitizing(element); // frame, editor

  element.style.display = "none";
  checkSanitizing(element); // no frame, editor

  form.removeChild(element);
}

for (type of todoTypes) {
  // The only meaning of this is to have a failure when new types are introduced
  // so we will know we have to write the tests here.
  var form = document.forms[0];
  var element = document.createElement("input");
  element.type = type;
  form.appendChild(element);

  todo_is(element.type, type, "");

  form.removeChild(element);
}

SpecialPowers.setBoolPref("dom.experimental_forms", pref);

</script>
</pre>
</body>
</html>
